requirements.txt
django
python-decouple
uagents
mailjet_rest
cs50